SE-301 renamed dummy files in unit tests for purpose clarity
