+++ sdk/features/unit/dummyfiles/duplicates/control_duplicate_decorator_values_dummy.py
@@ -3,16 +3,11 @@ import unittest
 from features.lusid_feature import lusid_feature
 
 
-class Valid2Tests(unittest.TestCase):
+class ThisClassContainsSameFeatureAsAnotherFileInThisPackage(unittest.TestCase):
 
-    @lusid_feature("F4")
+    @lusid_feature("F1")
     def test_dummy_method_1(self):
-        pass  # Empty for testing purposes
-
-    @lusid_feature("F3")
-    def test_dummy_method_2(self):
         pass  # Empty for testing purposes
 
     def test_control_method(self):
         pass  # Empty for testing purposes
-
