+++ sdk/features/unit/dummyfiles/duplicates/duplicate_decorator_values_accross_files_in_package.py
@@ -3,7 +3,7 @@ import unittest
 from features.lusid_feature import lusid_feature
 
 
-class DuplicateTests(unittest.TestCase):
+class ThisPackageContainsDuplicateFeatureAcrossTwoFiles(unittest.TestCase):
 
     @lusid_feature("F1")
     def test_dummy_method_1(self):
