+++ sdk/features/unit/dummyfiles/empties/control_file_with_empty_string_decorators.py
@@ -3,12 +3,12 @@ import unittest
 from features.lusid_feature import lusid_feature
 
 
-class NoinputTest(unittest.TestCase):
+class ThisClassAlsoContainsEmptyStringDecoratorValueAsControl(unittest.TestCase):
 
-    @lusid_feature()
+    @lusid_feature("")
     def test_dummy_method_1(self):
         pass  # Empty for testing purposes
 
-    @lusid_feature("F2")
+    @lusid_feature("F3")
     def test_control_method(self):
         pass  # Empty for testing purposes
