+++ sdk/features/unit/dummyfiles/empties/methods_containing_empty_decorator_values.py
@@ -1,10 +1,14 @@
 import unittest
 
+from features.lusid_feature import lusid_feature
 
-class NotAnnotatedTests(unittest.TestCase):
 
+class ThisClassContainsAnEmptyStringDecoratorValue(unittest.TestCase):
+
+    @lusid_feature("")
     def test_dummy_method_1(self):
         pass  # Empty for testing purposes
 
+    @lusid_feature("F2")
     def test_control_method(self):
         pass  # Empty for testing purposes
