+++ sdk/features/unit/dummyfiles/functions/decorated_functions_instead_of_classes_dummy.py
