+++ sdk/features/unit/dummyfiles/multidecorator/methods_containing_more_than_one_decorator.py
@@ -5,7 +5,7 @@ from parameterized import parameterized
 from features.lusid_feature import lusid_feature
 
 
-class Valid2Tests(unittest.TestCase):
+class ClassWithMethodsWithMoreThanOneDecorator(unittest.TestCase):
 
     @lusid_feature("F1")
     def test_dummy_method_1(self):
