+++ sdk/features/unit/dummyfiles/nobrackets/decorators_with_no_brackets_nor_values.py
