+++ sdk/features/unit/dummyfiles/noinput/class_with_decorator_with_brackets_but_no_input_value.py
@@ -1,10 +1,14 @@
 import unittest
 
+from features.lusid_feature import lusid_feature
 
-class NotAnnotated2Tests(unittest.TestCase):
 
+class ClassWithDecoratorWithBracketsButNoInputValue(unittest.TestCase):
+
+    @lusid_feature()
     def test_dummy_method_1(self):
         pass  # Empty for testing purposes
 
+    @lusid_feature("F2")
     def test_control_method(self):
         pass  # Empty for testing purposes
