+++ sdk/features/unit/dummyfiles/noinput/control_class_with_no_input_value_decorators.py
@@ -1,17 +1,14 @@
 import unittest
+
 from features.lusid_feature import lusid_feature
 
 
-class ValidTests(unittest.TestCase):
+class ControlClassWithDecoratorWithBracketsButNoInputValue(unittest.TestCase):
 
-    @lusid_feature("F1")
+    @lusid_feature()
     def test_dummy_method_1(self):
         pass  # Empty for testing purposes
 
-    @lusid_feature("F2")
-    def test_dummy_method_2(self):
-        pass  # Empty for testing purposes
-
+    @lusid_feature("F5")
     def test_control_method(self):
         pass  # Empty for testing purposes
-
