+++ sdk/features/unit/dummyfiles/notannotated/class_with_no_decorators_on_methods.py
@@ -1,14 +1,10 @@
 import unittest
 
-from features.lusid_feature import lusid_feature
 
+class ClassWithNoDecoratorsOnMethods(unittest.TestCase):
 
-class Empties2Tests(unittest.TestCase):
-
-    @lusid_feature("")
     def test_dummy_method_1(self):
         pass  # Empty for testing purposes
 
-    @lusid_feature("F3")
     def test_control_method(self):
         pass  # Empty for testing purposes
