+++ sdk/features/unit/dummyfiles/notannotated/control_class_with_no_decorators_on_methods.py
@@ -1,14 +1,10 @@
 import unittest
 
-from features.lusid_feature import lusid_feature
 
+class ControlClassWithNoDecoratorsOnMethods(unittest.TestCase):
 
-class EmptiesTests(unittest.TestCase):
-
-    @lusid_feature("")
     def test_dummy_method_1(self):
         pass  # Empty for testing purposes
 
-    @lusid_feature("F2")
     def test_control_method(self):
         pass  # Empty for testing purposes
