+++ sdk/features/unit/dummyfiles/valid/another_class_with_valid_decorators_on_methods.py
@@ -1,13 +1,17 @@
 import unittest
-
 from features.lusid_feature import lusid_feature
 
 
-class Duplicate2Tests(unittest.TestCase):
+class AnotherClassWithValidDecoratorsOnMethods(unittest.TestCase):
 
     @lusid_feature("F1")
     def test_dummy_method_1(self):
         pass  # Empty for testing purposes
 
+    @lusid_feature("F2")
+    def test_dummy_method_2(self):
+        pass  # Empty for testing purposes
+
     def test_control_method(self):
         pass  # Empty for testing purposes
+
